requirements.txt
asgiref==3.5.2
attrs==22.1.0
autopep8==2.0.0
certifi==2022.5.18.1
charset-normalizer==2.0.12
DateTime==5.0
decorator==5.1.1
Django==4.0.5
django-environ==0.8.1
djongo==1.3.6
dnspython==2.2.1
elementpath==3.0.2
future==0.18.3
idna==3.3
importlib-resources==5.10.0
isodate==0.6.1
jsonschema==4.16.0
jsonschema-spec==0.1.2
lazy-object-proxy==1.7.1
lxml==4.9.0
mongoengine==0.24.2
mongomock==4.1.2
openapi-schema-validator==0.3.4
openapi-spec-validator==0.5.1
packaging==21.3
pathable==0.4.3
pycodestyle==2.10.0
pyhandle==1.2.1
pymongo==3.12.3
PyMySQL==0.8.0
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
rdflib==6.1.1
requests==2.28.1
sentinels==1.0.0
six==1.16.0
sqlparse==0.2.4
tomli==2.0.1
typing_extensions==4.4.0
tzdata==2022.1
urllib3==1.26.9
validators==0.20.0
xmlschema==2.1.1
xmltodict==0.13.0
zope.interface==5.5.2
